Fix world state deserialization issue. Fixed radius for rects.

+++ DynmapClient/API/WorldState.cs
@@ -76,7 +76,7 @@
         public string Name { get; set; }
 
         [JsonProperty("timestamp")]
-        public int Timestamp { get; set; }
+        public long Timestamp { get; set; }
 
         [JsonProperty("msg", NullValueHandling = NullValueHandling.Ignore)]
         public string Msg { get; set; }
